+++ src/simmate/calculators/vasp/workflows/relaxation/all.py
@@ -2,6 +2,7 @@
 
 from .matproj import Relaxation__Vasp__Matproj
 from .matproj_hse import Relaxation__Vasp__MatprojHse
+from .matproj_scan import Relaxation__Vasp__MatprojScan
 from .mit import Relaxation__Vasp__Mit
 from .neb_endpoint import Relaxation__Vasp__NebEndpoint
 from .quality00 import Relaxation__Vasp__Quality00