+++ src/simmate/workflows/static_energy.py
@@ -9,6 +9,7 @@ Workflows for calculating the energy of material
 from simmate.calculators.vasp.workflows.static_energy.all import (
     StaticEnergy__Vasp__Matproj,
     StaticEnergy__Vasp__MatprojHse,
+    StaticEnergy__Vasp__MatprojScan,
     StaticEnergy__Vasp__Mit,
     StaticEnergy__Vasp__Quality04,
     StaticEnergy__Vasp__NebEndpoint,